+++ src/mkvtoolnix-gui/chapter_editor/chapter_model.cpp
@@ -182,9 +182,9 @@ ChapterModel::chapterNameForLanguage(KaxChapterAtom &chapter,
     if (!kDisplay)
       continue;
 
-    auto lists = NameModel::effectiveLanguagesAndCountriesForDisplay(*kDisplay);
-
+    auto lists          = NameModel::effectiveLanguagesForDisplay(*kDisplay);
     auto actualLanguage = mtx::chapters::get_language_from_display(*kDisplay, "eng"s);
+
     if (   language.empty()
         || (std::find_if(lists.languageCodes.begin(), lists.languageCodes.end(), [&language](auto const &actualLanguage) {
               return (language == actualLanguage.get_language())